Crate hacspec_riot_runqueue

Structs

Clist

NextIds
Fixed length byte array.

RunQueue

RunqueueId

Tail
Fixed length byte array.

ThreadId

Functions

clist_advance

clist_is_empty

clist_new

clist_peek_head

clist_pop_head

clist_push

runqueue_add

runqueue_advance
advance runqueue number rq (this is used to "yield" to another thread of the same priority)

runqueue_del

runqueue_ffs

runqueue_get_next
get pid that should run next returns the next runnable thread of the runqueue with the highest index

runqueue_new